ignore removed deps in fsckprojects

+++ mx.py
@@ -15120,16 +15120,18 @@ def fsckprojects(args):
                 # don't traverse subdirs of an existing distribution in this suite
                 dirnames[:] = []
             else:
-                projectConfigFiles = frozenset(['.classpath', '.project', 'nbproject', basename(dirpath) + '.iml'])
-                indicators = projectConfigFiles.intersection(files)
-                if len(indicators) != 0:
-                    indicators = [os.path.relpath(join(dirpath, i), suite.vc_dir) for i in indicators]
-                    indicatorsInVC = suite.vc.locate(suite.vc_dir, indicators)
-                    # Only proceed if there are indicator files that are not under VC
-                    if len(indicators) > len(indicatorsInVC):
-                        if ask_yes_no(dirpath + ' looks like a removed project -- delete it', 'n'):
-                            shutil.rmtree(dirpath)
-                            log('Deleted ' + dirpath)
+                maybe_project = basename(dirpath)
+                if not _removedDeps.get(maybe_project):
+                    projectConfigFiles = frozenset(['.classpath', '.project', 'nbproject', maybe_project + '.iml'])
+                    indicators = projectConfigFiles.intersection(files)
+                    if len(indicators) != 0:
+                        indicators = [os.path.relpath(join(dirpath, i), suite.vc_dir) for i in indicators]
+                        indicatorsInVC = suite.vc.locate(suite.vc_dir, indicators)
+                        # Only proceed if there are indicator files that are not under VC
+                        if len(indicators) > len(indicatorsInVC):
+                            if ask_yes_no(dirpath + ' looks like a removed project -- delete it', 'n'):
+                                shutil.rmtree(dirpath)
+                                log('Deleted ' + dirpath)
         ideaProjectDirectory = join(suite.dir, '.idea')
         librariesDirectory = join(ideaProjectDirectory, 'libraries')
         if exists(librariesDirectory):
